external/tina/requirements.txt
numpy
pytest
taichi
gltflib
pillow
pysobol
